added logical room switching
the code now sets the room to the room with the staircase on for the next floor.
Also moved the changefloor function to the player class, where it belongs.
Also added the framework for leaving buildings, just have to figure out the input/logic for when to execute the code.